+++ tests/test_processor.py
@@ -80,23 +80,6 @@ class NifiProcessor(unittest.TestCase):
             body={'revision': {'clientId': 'mock_processor_id', 'version': 1}, 'state': 'RUNNING'}
         )
 
-    @patch('nipyapi.nifi.apis.processors_api.ProcessorsApi.get_processor')
-    @patch('nipyapi.nifi.apis.processors_api.ProcessorsApi.update_run_status')
-    def test_stop_processor_stop_processor_successfully(self, mock_update_run_status, mock_get_processor):
-        mock_response = MagicMock()
-        mock_response.revision.version = 1
-        mock_response.id = 'mock_processor_id'
-        mock_get_processor.return_value = mock_response
-        mock_update_run_status.return_value = 'Success'
-
-        result = self.nifi.stop_processor(self.processor_id, return_response=True)
-
-        assert result == 'Success'
-        mock_update_run_status.assert_called_once_with(
-            id='mock_processor_id',
-            body={'revision': {'clientId': 'mock_processor_id', 'version': 1}, 'state': 'STOPPED'}
-        )
-
     def test_stop_processor_raises_exception_for_missing_processor_id(self):
         try:
             self.nifi.stop_processor(None)
@@ -118,14 +101,14 @@ class NifiProcessor(unittest.TestCase):
         except Exception as e:
             assert str(e) == 'Failed to stop processor'
 
-    # @patch('NifiLibrary.NifiLibrary.update_process_state')
-    # def test_start_processor_starts_processor_successfully(self, mock_update_process_state):
-    #     mock_update_process_state.return_value = 'Success'
-    #
-    #     result = self.nifi.start_processor(self.processor_id, return_response=True)
-    #
-    #     assert result == 'Success'
-    #     mock_update_process_state.assert_called_once_with(self.processor_id, 'RUNNING')
+    @patch('NifiLibrary.NifiLibrary.update_process_state')
+    def test_start_processor_starts_processor_successfully(self, mock_update_process_state):
+        mock_update_process_state.return_value = 'Success'
+
+        result = self.nifi.start_processor(self.processor_id, return_response=True)
+
+        assert result == 'Success'
+        mock_update_process_state.assert_called_once_with(self.processor_id, 'RUNNING')
 
     def test_start_processor_raises_exception_for_missing_processor_id(self):
         try:
@@ -133,40 +116,30 @@ class NifiProcessor(unittest.TestCase):
         except Exception as e:
             assert str(e) == 'Require parameters cannot be none'
 
-    # @patch('NifiLibrary.NifiLibrary.NifiLibrary.update_process_state')
-    # def test_start_processor_logs_error_on_failure(self, mock_update_process_state):
-    #     mock_update_process_state.side_effect = Exception('Failed to start processor')
-    #
-    #     try:
-    #         self.nifi.start_processor(self.processor_id)
-    #     except Exception as e:
-    #         assert str(e) == 'Failed to start processor'
-    #     mock_update_process_state.assert_called_once_with(self.processor_id, 'RUNNING')
-    #
-    # # @patch('NifiLibrary.NifiLibrary.NifiLibrary.update_process_state')
-    # # def test_run_once_processor_runs_successfully(self, mock_update_process_state):
-    # #     mock_update_process_state.return_value = 'Success'
-    # #
-    # #     result = self.nifi.run_once_processor(self.processor_id, return_response=True)
-    # #
-    # #     assert result == 'Success'
-    # #     mock_update_process_state.assert_called_once_with(self.processor_id, 'RUN_ONCE')
-    #
-    # def test_run_once_processor_raises_exception_for_missing_processor_id(self):
-    #     try:
-    #         self.nifi.run_once_processor(None)
-    #     except Exception as e:
-    #         assert str(e) == 'Require parameters cannot be none'
-    #
-    # # @patch('NifiLibrary.NifiLibrary.NifiLibrary.update_process_state')
-    # # def test_run_once_processor_logs_error_on_failure(self, mock_update_process_state):
-    # #     mock_update_process_state.side_effect = Exception('Failed to run processor once')
-    # #
-    # #     try:
-    # #         self.nifi.run_once_processor(self.processor_id)
-    # #     except Exception as e:
-    # #         assert str(e) == 'Failed to run processor once'
-    # #     mock_update_process_state.assert_called_once_with(self.processor_id, 'RUN_ONCE')
-    # #
+    @patch('NifiLibrary.NifiLibrary.update_process_state')
+    def test_start_processor_failure(self, mock_update_process_state):
+        mock_update_process_state.side_effect = Exception('Failed to start processor')
+        with self.assertRaisesRegex(Exception, 'Failed to start processor'):
+            self.nifi.start_processor(self.processor_id)
+        mock_update_process_state.assert_called_once_with(self.processor_id, 'RUNNING')
+
+    @patch('NifiLibrary.NifiLibrary.update_run_once_process_state')
+    def test_run_once_processor_success(self, mock_update_process_state):
+        mock_update_process_state.return_value = 'Success'
+        result = self.nifi.run_once_processor(self.processor_id, return_response=True)
+        assert result == 'Success'
+        mock_update_process_state.assert_called_once_with(self.processor_id, 'RUN_ONCE')
+
+    def test_run_once_processor_missing_processor_id(self):
+        with self.assertRaisesRegex(Exception, 'Require parameters cannot be none'):
+            self.nifi.run_once_processor(None)
+
+    @patch('NifiLibrary.NifiLibrary.update_run_once_process_state')
+    def test_run_once_processor_failure(self, mock_update_process_state):
+        mock_update_process_state.side_effect = Exception('Failed to run processor once')
+        with self.assertRaisesRegex(Exception, 'Failed to run processor once'):
+            self.nifi.run_once_processor(self.processor_id)
+        mock_update_process_state.assert_called_once_with(self.processor_id, 'RUN_ONCE')
+
     if __name__ == '__main__':
         unittest.main()